requirements.txt
Flask==2.3.3
requests
python-dotenv